Simplify camera code
Prior to this change, the camera code used much more lines of code.

Also, the attacks would rotate with the player. It has been temporarily
removed.

This change makes the code simpler and prepares it for the code
necessary for aiming of attacks.

The simpler camera code comes from this video: https://youtu.be/sNmeK3qK7oA